refactor elevator logic out into its own class to keep things cleaner
